backend/requirements.txt
flask
flask-cors
nvdlib
requests
gitpython
